Feb 02 2017

+++ Project1/ActorInfo.cpp
@@ -1,6 +1,6 @@
 #include "ActorInfo.h"
 
-ActorInfo::ActorInfo(int x, int y, int type, int life)
+ActorInfo::ActorInfo(float x, float y, int type, int life)
 {
 	initialize(x, y, type, life);
 }
@@ -14,7 +14,7 @@ ActorInfo::~ActorInfo()
 {
 }
 
-void ActorInfo::initialize(int x, int y, int type, int life)
+void ActorInfo::initialize(float x, float y, int type, int life)
 {
 	_pt.x = x;
 	_pt.y = y;
